fix(phase-12b): correct BIGINT→DOUBLE PRECISION for score_snapshots.n_obs
Finding A of the Phase 12B migration audit: `score_snapshots.n_obs` was
ported from SQLite (permissive INTEGER) to Postgres as BIGINT, but the
column actually stores `nObsEffective = (α + β) − (α₀ + β₀)` — a decayed
real-valued weight produced by `bayesianVerdictService.buildVerdict`
(round3 of `combined.nObs`), not a raw observation counter.

Under strict Postgres typing, every rescore attempt emitted
`invalid input syntax for type bigint: "0.987"` and the snapshot insert
failed silently, leaving `unscoredCount` stuck and blocking new
score_snapshots rows for any agent with decayed evidence.

Fix scope is limited to this one column. Audit of all bayesian tables
(score_snapshots, *_streaming_posteriors ×5, *_daily_buckets ×5,
nostr_published_events) confirmed no other column is mistyped. In
particular `nostr_published_events.n_obs_effective DOUBLE PRECISION`
already has the correct type for the exact same semantic — the Postgres
port had the right pattern for the Nostr ledger but missed it for
score_snapshots. `total_ingestions` stays BIGINT (raw +1 counter,
confirmed by `streamingPosteriorRepository.ts:165` and MIN=MAX=1 in
prod). `*_daily_buckets.n_obs` stays BIGINT (daily integer counter).

Changes:
- ALTER TABLE score_snapshots ALTER COLUMN n_obs TYPE DOUBLE PRECISION
  executed on prod in 128.7 ms. The 12,291 pre-existing rows all had
  n_obs = 0 (legacy SQLite pre-streaming), so the cast is lossless.
- src/database/postgres-schema.sql: keep the consolidated schema in
  sync so fresh installs (and the vitest template DB) get the correct
  type from the start.
- src/tests/snapshotNobsFloat.test.ts: regression test covering the
  canonical failing value 0.987 plus boundary cases (0, 42, 12.375,
  1_000_000.125).

Post-fix verification: one bulk rescore cycle wrote 5,515 new snapshots
with real float n_obs (max observed 0.982). Zero bigint errors over the
following 5 minutes of crawler logs. Four of the five previously
reported blocked agents (fa44376c, cb0c2aff, ec1c4124, f35ed6ba) now
have fresh snapshots; the fifth (6bea5652) is pending the next cycle
with no specific error.

+++ src/database/postgres-schema.sql
@@ -90,7 +90,7 @@ CREATE TABLE IF NOT EXISTS score_snapshots (
   p_success         DOUBLE PRECISION,
   ci95_low          DOUBLE PRECISION,
   ci95_high         DOUBLE PRECISION,
-  n_obs             BIGINT,
+  n_obs             DOUBLE PRECISION,
   "window"          TEXT,
   updated_at        BIGINT
 );
